+++ kohii-core/src/main/java/kohii/v1/core/AbstractPlayable.kt
@@ -22,7 +22,6 @@ import kohii.v1.core.MemoryMode.HIGH
 import kohii.v1.core.MemoryMode.INFINITE
 import kohii.v1.core.MemoryMode.LOW
 import kohii.v1.core.MemoryMode.NORMAL
-import kohii.v1.core.Playback.Callback
 import kohii.v1.internal.PlayerParametersChangeListener
 import kohii.v1.logInfo
 import kohii.v1.logWarn
@@ -35,7 +34,7 @@ abstract class AbstractPlayable<RENDERER : Any>(
   media: Media,
   config: Config,
   protected val bridge: Bridge<RENDERER>
-) : Playable(media, config), Callback, PlayerParametersChangeListener {
+) : Playable(media, config), Playback.Callback, PlayerParametersChangeListener {
 
   override val tag: Any = config.tag
 
@@ -98,18 +97,18 @@ abstract class AbstractPlayable<RENDERER : Any>(
 
   override var manager: PlayableManager? = null
     set(value) {
-      val from = field
+      val oldManager = field
       field = value
-      val to = field
-      if (from === to) return
-      "Playable#manager $from --> $to, $this".logInfo()
-      if (to == null) {
+      val newManager = field
+      if (oldManager === newManager) return
+      "Playable#manager $oldManager --> $newManager, $this".logInfo()
+      if (newManager == null) {
         master.trySavePlaybackInfo(this)
         master.tearDown(
             playable = this,
-            clearState = if (from is Manager) !from.isChangingConfigurations() else true
+            clearState = if (oldManager is Manager) !oldManager.isChangingConfigurations() else true
         )
-      } else if (from === null) {
+      } else if (oldManager === null) {
         master.tryRestorePlaybackInfo(this)
       }
     }
@@ -134,17 +133,19 @@ abstract class AbstractPlayable<RENDERER : Any>(
       this.manager = if (newPlayback != null) {
         newPlayback.manager
       } else {
-        val configChange = oldPlayback?.manager?.isChangingConfigurations() == true
-        if (!configChange) {
+        val changingConfiguration = oldPlayback?.manager?.isChangingConfigurations() == true
+        if (changingConfiguration) {
+          if (!onConfigChange()) {
+            // On config change, if the Playable doesn't support, we need to pause the Video.
+            onPause()
+            null
+          } else {
+            master // to prevent the Playable from being destroyed when Manager is null.
+          }
+        } else {
           // TODO need a better implementation.
           if (master.manuallyStartedPlayable.get() === this && isPlaying()) master
           else null
-        } else if (!onConfigChange()) {
-          // On config change, if the Playable doesn't support, we need to pause the Video.
-          onPause()
-          null
-        } else {
-          master // to prevent the Playable from being destroyed when Manager is null.
         }
       }
 
@@ -186,7 +187,7 @@ abstract class AbstractPlayable<RENDERER : Any>(
     "Playable#onInActive $playback, $this".logInfo()
     require(playback === this.playback)
     val configChange = playback.manager.isChangingConfigurations()
-    if (!configChange && !master.onPlaybackInActive(this, playback)) {
+    if (!configChange && master.releasePlaybackOnInActive(playback)) {
       master.trySavePlaybackInfo(this)
       master.releasePlayable(this)
     }
@@ -210,18 +211,25 @@ abstract class AbstractPlayable<RENDERER : Any>(
     master.onPlaybackDetached(playback)
   }
 
-  override fun considerRequestRenderer(playback: Playback) {
-    "Playable#considerRequestRenderer $playback, $this".logInfo()
+  override fun setupRenderer(playback: Playback) {
+    "Playable#setupRenderer $playback, $this".logInfo()
     require(playback === this.playback)
     if (this.renderer == null || manager !== playback.manager) {
       // Only request for Renderer if we do not have one.
       val renderer = playback.acquireRenderer()
-      if (playback.attachRenderer(renderer)) this.renderer = renderer
+      if (playback.attachRenderer(renderer)) {
+        this.renderer = renderer
+        onRendererAttached(playback, renderer)
+      }
     }
   }
 
-  override fun considerReleaseRenderer(playback: Playback) {
-    "Playable#considerReleaseRenderer $playback, $this".logInfo()
+  protected open fun onRendererAttached(playback: Playback, renderer: Any?) {
+    playback.onRendererAttached(renderer)
+  }
+
+  override fun teardownRenderer(playback: Playback) {
+    "Playable#teardownRenderer $playback, $this".logInfo()
     require(this.playback == null || this.playback === playback)
     val renderer = this.renderer
     if (renderer != null) {
@@ -229,30 +237,34 @@ abstract class AbstractPlayable<RENDERER : Any>(
       if (playback.detachRenderer(renderer)) {
         playback.releaseRenderer(renderer)
         this.renderer = null
+        onRendererDetached(playback, renderer)
       }
     }
   }
 
-  override fun onDistanceChanged(
+  protected open fun onRendererDetached(playback: Playback, renderer: Any?) {
+    playback.onRendererDetached(renderer)
+  }
+
+  override fun onPlaybackPriorityChanged(
     playback: Playback,
-    from: Int,
-    to: Int
+    oldPriority: Int,
+    newPriority: Int
   ) {
-    "Playable#onDistanceChanged $playback, $from --> $to, $this".logInfo()
-    if (to == 0) {
+    "Playable#onPlaybackPriorityChanged $playback, $oldPriority --> $newPriority, $this".logInfo()
+    if (newPriority == 0) {
       master.tryRestorePlaybackInfo(this)
       master.preparePlayable(this, playback.config.preload)
     } else {
       val memoryMode = master.preferredMemoryMode(this.memoryMode)
-      val distanceToRelease =
-        when (memoryMode) {
-          AUTO, LOW -> 1
-          NORMAL -> 2
-          BALANCED -> 2 // Same as 'NORMAL', but will keep the 'relative' Playback alive.
-          HIGH -> 8
-          INFINITE -> Int.MAX_VALUE - 1
-        }
-      if (to >= distanceToRelease) {
+      val priorityToRelease = when (memoryMode) {
+        AUTO, LOW -> 1
+        NORMAL -> 2
+        BALANCED -> 2 // Same as 'NORMAL', but will keep the 'relative' Playback alive.
+        HIGH -> 8
+        INFINITE -> Int.MAX_VALUE - 1
+      }
+      if (newPriority >= priorityToRelease) {
         master.trySavePlaybackInfo(this)
         master.releasePlayable(this)
       } else {